split up js and wasm
